+++ cognite/neat/_rules/transformers/_converters.py
@@ -4,7 +4,7 @@ from abc import ABC, abstractmethod
 from collections import Counter, defaultdict
 from collections.abc import Collection, Mapping
 from datetime import date, datetime
-from typing import Literal, TypeVar, cast
+from typing import Literal, TypeVar, cast, overload
 
 from cognite.client.data_classes import data_modeling as dms
 from cognite.client.data_classes.data_modeling import DataModelId, DataModelIdentifier, ViewId
@@ -42,8 +42,10 @@ from cognite.neat._rules.models.entities import (
     ContainerEntity,
     DMSUnknownEntity,
     EdgeEntity,
+    Entity,
     MultiValueTypeInfo,
     ReverseConnectionEntity,
+    T_Entity,
     UnknownEntity,
     ViewEntity,
 )
@@ -174,6 +176,74 @@ class ToCompliantEntities(RulesTransformer[InformationInputRules, InformationInp
             definition.value_type = cls._fix_value_type(definition.value_type)
             fixed_definitions.append(definition)
         return fixed_definitions
+
+
+class PrefixEntities(RulesTransformer[InputRules, InputRules]):  # type: ignore[misc]
+    """Prefixes all entities with a given prefix."""
+
+    def __init__(self, prefix: str) -> None:
+        self._prefix = prefix
+
+    def transform(self, rules: InputRules | OutRules[InputRules]) -> ReadRules[InputRules]:
+        return ReadRules(self._transform(self._to_rules(rules)), IssueList(), {})
+
+    def _transform(self, rules: InputRules) -> InputRules:
+        rules.metadata.version += f"_prefixed_{self._prefix}"
+
+        if isinstance(rules, InformationInputRules):
+            # Todo Make Not mutate input class
+            prefixed_by_class: dict[str, str] = {}
+            for cls in rules.classes:
+                prefixed = str(self._with_prefix(cls.class_))
+                prefixed_by_class[str(cls.class_)] = prefixed
+                cls.class_ = prefixed
+            for prop in rules.properties:
+                prop.class_ = self._with_prefix(prop.class_)
+                if str(prop.value_type) in prefixed_by_class:
+                    prop.value_type = prefixed_by_class[str(prop.value_type)]
+            return rules
+        elif isinstance(rules, DMSInputRules):
+            # Todo not mutate input class new_dms = copy.deepcopy(rules)
+            prefixed_by_view: dict[str, str] = {}
+            for view in rules.views:
+                prefixed = str(self._with_prefix(view.view))
+                prefixed_by_view[str(view.view)] = prefixed
+                view.view = prefixed
+            for dms_prop in rules.properties:
+                dms_prop.view = self._with_prefix(dms_prop.view)
+                if str(dms_prop.value_type) in prefixed_by_view:
+                    dms_prop.value_type = prefixed_by_view[str(dms_prop.value_type)]
+            if rules.containers:
+                for container in rules.containers:
+                    container.container = self._with_prefix(container.container)
+            return rules
+        raise NeatValueError(f"Unsupported rules type: {type(rules)}")
+
+    @overload
+    def _with_prefix(self, raw: str) -> str: ...
+
+    @overload
+    def _with_prefix(self, raw: T_Entity) -> T_Entity: ...
+
+    def _with_prefix(self, raw: str | T_Entity) -> str | T_Entity:
+        is_entity_format = not isinstance(raw, str)
+        entity = Entity.load(raw)
+        output: ClassEntity | ViewEntity | ContainerEntity
+        if isinstance(entity, ClassEntity):
+            output = ClassEntity(prefix=entity.prefix, suffix=f"{self._prefix}{entity.suffix}", version=entity.version)
+        elif isinstance(entity, ViewEntity):
+            output = ViewEntity(
+                space=entity.space, externalId=f"{self._prefix}{entity.external_id}", version=entity.version
+            )
+        elif isinstance(entity, ContainerEntity):
+            output = ContainerEntity(space=entity.space, externalId=f"{self._prefix}{entity.external_id}")
+        elif isinstance(entity, UnknownEntity | Entity):
+            return f"{self._prefix}{raw}"
+        else:
+            raise NeatValueError(f"Unsupported entity type: {type(entity)}")
+        if is_entity_format:
+            return cast(T_Entity, output)
+        return str(output)
 
 
 class InformationToDMS(ConversionTransformer[InformationRules, DMSRules]):