+++ docs/CHANGELOG.md
@@ -18,6 +18,7 @@ Changes are grouped as follows:
 ## TBD
 ### Added
 - Support for dropping instances of a given type with `neat.drop.instances(...)`
+- Helper methods `neat.prepare.instances.relationships_as_connection` and `neat.prepare.data_model.prefix` added
 
 ### Fixed
 - Running `neat.to.cdf.data_model(..., dry_run=True)` now counts created and deleted items correctly.
@@ -27,6 +28,7 @@ Changes are grouped as follows:
 - Support for reading `YAML`
 - Support for writing `YAML` in toolkit format.
 - Added end-to-end test for CDM extension
+- Support for dropping instances of a given type with `neat.drop.instances(...)`
 ### Fixed
 - Verification of reversed connections was done of `property (linage)` instead of `view_property`
 ### Changed